Fix
Cast incoming timestamp (int) to int64

+++ function/air/epoch2timestamp.go
@@ -25,7 +25,7 @@ func (fnEpoch2Timestamp) Sig() (paramTypes []data.Type, isVariadic bool) {
 func (fnEpoch2Timestamp) Eval(params ...interface{}) (interface{}, error) {
 	/* quick and dirty : only for realtime scenario */
 	log.Info("(fnEpoch2Timestamp.Eval) params[0] : ", params[0], ", params[1] : ", params[1])
-	epoch := params[0].(int64)
+	epoch := int64(params[0].(int))
 	if epoch > 1000000000000000000 {
 		epoch = epoch / 1000000000
 	} else if epoch > 1000000000000 {
